Ejemplo resuelto

+++ src/com/tp/ia/src/CambioMinimoFuncionAptitud.java
@@ -6,16 +6,16 @@ import org.jgap.IChromosome;
 public class CambioMinimoFuncionAptitud extends FitnessFunction {
 
     private final int montoObjetivo;
-// Maximo monto posible 1000 Centimos = 10 Euros
+    // Maximo monto posible 1000 Centimos = 10 Euros
     public static final int MAX_MONTO = 1000;
-// Maxima cantidad de monedas posibles. Es igual al Monto maximo en
+    // Maxima cantidad de monedas posibles. Es igual al Monto maximo en
 
     // centimos, ya que si se utilizan monedas de un centimo se llegaria al
-// monton con la mayor cantidad posible de monedas
+    // monton con la mayor cantidad posible de monedas
     public static final int MAX_CANT_MONEDAS = MAX_MONTO;
-// El constructor de la funcion de aptitud debe recibir el monto objetivo
-// del problema y almacenarlo en un atributo. Si el monto es invalido arroja
-// una excepcion
+    // El constructor de la funcion de aptitud debe recibir el monto objetivo
+    // del problema y almacenarlo en un atributo. Si el monto es invalido arroja
+    // una excepcion
 
     public CambioMinimoFuncionAptitud(int monto) {
         if (monto < 1 || monto >= MAX_MONTO) {
@@ -26,19 +26,19 @@ public class CambioMinimoFuncionAptitud extends FitnessFunction {
 
     @Override
     public double evaluate(IChromosome cromosoma) {
-// Se debe tener en cuenta el evaluador que se esta usando. El evaluador
-// estandar le asigna un valor mas apto a los valores mas altos de
-// aptitud. Tambien hay otros evaluadores que asignan mejor aptitud a
-// los valores mas bajos.
-// Es por esto que se chequea si 2 es mas apto que 1. Si esto es asi
-// entonces el valor mas apto sera el mayor y el menos apto el 0
+        // Se debe tener en cuenta el evaluador que se esta usando. El evaluador
+        // estandar le asigna un valor mas apto a los valores mas altos de
+        // aptitud. Tambien hay otros evaluadores que asignan mejor aptitud a
+        // los valores mas bajos.
+        // Es por esto que se chequea si 2 es mas apto que 1. Si esto es asi
+        // entonces el valor mas apto sera el mayor y el menos apto el 0
         boolean evaluadorEstandard = cromosoma.getConfiguration().getFitnessEvaluator().isFitter(2, 1);
         int montoCambioMonedas = montoCambioMoneda(cromosoma);
         int totalMonedas = getNumeroTotalMonedas(cromosoma);
         int diferenciaMonto = Math.abs(montoObjetivo - montoCambioMonedas);
-// El primer paso es asignar la menor aptitud a aquellos cromosomas cuyo
-// monto no sea el monto objetivo. Es decir una descomposicion en
-// monedas que no sea del monto ingresado
+        // El primer paso es asignar la menor aptitud a aquellos cromosomas cuyo
+        // monto no sea el monto objetivo. Es decir una descomposicion en
+        // monedas que no sea del monto ingresado
         if (evaluadorEstandard) {
             if (diferenciaMonto != 0) {
                 return 0.0d;
@@ -48,16 +48,16 @@ public class CambioMinimoFuncionAptitud extends FitnessFunction {
                 return MAX_CANT_MONEDAS;
             }
         }
-// luego se debe asignar mas aptitud a aquellos cromosomas que posean
-// menor cantidad de monedas.
+        // luego se debe asignar mas aptitud a aquellos cromosomas que posean
+        // menor cantidad de monedas.
         if (evaluadorEstandard) {
-// Se debe asegurar devolver un valor de aptitud positivo siempre.
-// Si el valor es negativo se devuelve MAX_CANT_MONEDAS ( elemento
-// menos apto )
+        // Se debe asegurar devolver un valor de aptitud positivo siempre.
+        // Si el valor es negativo se devuelve MAX_CANT_MONEDAS ( elemento
+        // menos apto )
             return Math.max(0.0d, MAX_CANT_MONEDAS - totalMonedas);
         } else {
-// Se debe asgurar devolver un valor de aptitud positivo siempre.
-// Si el valor es negativo se devuelve 0 ( elemento menos apto )
+        // Se debe asgurar devolver un valor de aptitud positivo siempre.
+        // Si el valor es negativo se devuelve 0 ( elemento menos apto )
             return Math.max(0.0d, totalMonedas);
         }
     }
@@ -74,7 +74,7 @@ public class CambioMinimoFuncionAptitud extends FitnessFunction {
 
     public static int getNumeroDeComendasDeGen(IChromosome cromosoma, int numeroGen) {
         Integer numMonedas = (Integer) cromosoma.getGene(numeroGen).getAllele();
-        return numMonedas.intValue();
+        return numMonedas;
     }
 
     public static int getNumeroTotalMonedas(IChromosome cromosoma) {
